refactor(PlayService, ResultService): Service 클래스 분리
