+++ README.md
@@ -100,5 +100,5 @@ erDiagram
 - [x] 예측하지 못한 예외도 핸들러로 관리
 - [x] SimpleJdbcInsert 사용
 - [x] 사용하지 않는 주석 제거
-- [ ] 의미있는 클래스 명으로 변경
+- [x] Service 클래스 분리
 - [x] service에서 domain으로 책임 이동
